UBERF-5837 Better organize document plugin
