+++ plugins/document-assets/lang/en.json
@@ -1,21 +1,20 @@
 {
   "string": {
+    "Document": "Document",
+    "ParentDocument": "Parent document",
+    "ChildDocument": "Child document",
     "CreateDocument": "Create a document",
     "Documents": "Documents",
     "CreateAnVersion": "Create new version",
-    "Document": "Document",
     "DocumentNamePlaceholder": "Untitled",
     "Name": "Name",
     "Cancel": "Cancel",
     "Version": "Version",
     "SearchDocument": "Search document...",
     "Snapshot": "Snapshot",
-    "ParentDocument": "Parent document",
-    "ChildDocument": "Child document",
     "NoVersions": "No version's defined",
     "DocumentApplication": "Documents",
     "MyDocuments": "My documents",
-    "Library": "Library",
     "Teamspace": "Teamspace",
     "Teamspaces": "Teamspaces",
     "Labels": "Labels",
@@ -56,7 +55,6 @@
     "UnassignToDoConfirm": "Do you want to unassign Todo? The Todo will be removed from the assignee planning.",
     "ReassignToDo": "Change Todo assignee",
     "ReassignToDoConfirm": "Do you want to change Todo assignee? The Todo will be removed from the current assignee's planning.",
-    "Docs": "Docs",
     "Icon": "Icon",
     "Color": "Color"
   }
