+++ plugins/document-resources/src/components/teamspace/TeamspacePresenter.svelte
@@ -27,7 +27,7 @@
   <div class="flex-presenter cursor-default" class:inline-presenter={inline}>
     <div class="icon" class:emoji={value.icon === view.ids.IconWithEmoji}>
       <Icon
-        icon={value.icon === view.ids.IconWithEmoji ? IconWithEmoji : value.icon ?? document.icon.Library}
+        icon={value.icon === view.ids.IconWithEmoji ? IconWithEmoji : value.icon ?? document.icon.Teamspace}
         iconProps={value.icon === view.ids.IconWithEmoji
           ? { icon: value.color }
           : {
